requirements.txt
anthropic
requests
python-dotenv
gspread
google-auth
google-api-python-client
